move: OAuthProvider utils -> domain
