+++ src/main/java/com/kgu/studywithme/auth/application/adapter/OAuthUriGenerator.java
@@ -1,6 +1,6 @@
 package com.kgu.studywithme.auth.application.adapter;
 
-import com.kgu.studywithme.auth.utils.OAuthProvider;
+import com.kgu.studywithme.auth.domain.oauth.OAuthProvider;
 
 public interface OAuthUriGenerator {
     boolean isSupported(final OAuthProvider provider);
