+++ src/main/java/com/kgu/studywithme/auth/application/service/OAuthLoginService.java
@@ -5,11 +5,11 @@ import com.kgu.studywithme.auth.application.adapter.TokenPersistenceAdapter;
 import com.kgu.studywithme.auth.application.usecase.command.OAuthLoginUseCase;
 import com.kgu.studywithme.auth.domain.AuthMember;
 import com.kgu.studywithme.auth.domain.AuthToken;
+import com.kgu.studywithme.auth.domain.oauth.OAuthProvider;
 import com.kgu.studywithme.auth.domain.oauth.OAuthTokenResponse;
 import com.kgu.studywithme.auth.domain.oauth.OAuthUserResponse;
 import com.kgu.studywithme.auth.exception.AuthErrorCode;
 import com.kgu.studywithme.auth.utils.JwtTokenProvider;
-import com.kgu.studywithme.auth.utils.OAuthProvider;
 import com.kgu.studywithme.global.exception.StudyWithMeException;
 import com.kgu.studywithme.global.exception.StudyWithMeOAuthException;
 import com.kgu.studywithme.member.application.adapter.MemberReadAdapter;
