+++ src/main/java/com/kgu/studywithme/auth/application/usecase/query/QueryOAuthLinkUseCase.java
@@ -1,6 +1,6 @@
 package com.kgu.studywithme.auth.application.usecase.query;
 
-import com.kgu.studywithme.auth.utils.OAuthProvider;
+import com.kgu.studywithme.auth.domain.oauth.OAuthProvider;
 
 public interface QueryOAuthLinkUseCase {
     String invoke(final Query query);
