+++ src/main/java/com/kgu/studywithme/auth/infrastructure/oauth/naver/NaverOAuthConnector.java
@@ -1,12 +1,12 @@
 package com.kgu.studywithme.auth.infrastructure.oauth.naver;
 
 import com.kgu.studywithme.auth.application.adapter.OAuthConnector;
+import com.kgu.studywithme.auth.domain.oauth.OAuthProvider;
 import com.kgu.studywithme.auth.domain.oauth.OAuthTokenResponse;
 import com.kgu.studywithme.auth.domain.oauth.OAuthUserResponse;
 import com.kgu.studywithme.auth.exception.AuthErrorCode;
 import com.kgu.studywithme.auth.infrastructure.oauth.naver.response.NaverTokenResponse;
 import com.kgu.studywithme.auth.infrastructure.oauth.naver.response.NaverUserResponse;
-import com.kgu.studywithme.auth.utils.OAuthProvider;
 import com.kgu.studywithme.global.exception.StudyWithMeException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpEntity;
@@ -18,9 +18,9 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 
+import static com.kgu.studywithme.auth.domain.oauth.OAuthProvider.NAVER;
 import static com.kgu.studywithme.auth.infrastructure.oauth.OAuthMetadata.CONTENT_TYPE_VALUE;
 import static com.kgu.studywithme.auth.infrastructure.oauth.OAuthMetadata.TOKEN_TYPE;
-import static com.kgu.studywithme.auth.utils.OAuthProvider.NAVER;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 import static org.springframework.http.HttpMethod.GET;
