+++ src/main/java/com/kgu/studywithme/auth/presentation/OAuthApiController.java
@@ -4,9 +4,9 @@ import com.kgu.studywithme.auth.application.usecase.command.LogoutUseCase;
 import com.kgu.studywithme.auth.application.usecase.command.OAuthLoginUseCase;
 import com.kgu.studywithme.auth.application.usecase.query.QueryOAuthLinkUseCase;
 import com.kgu.studywithme.auth.domain.AuthMember;
+import com.kgu.studywithme.auth.domain.oauth.OAuthProvider;
 import com.kgu.studywithme.auth.presentation.dto.request.OAuthLoginRequest;
 import com.kgu.studywithme.auth.utils.ExtractPayload;
-import com.kgu.studywithme.auth.utils.OAuthProvider;
 import com.kgu.studywithme.global.aop.CheckAuthUser;
 import com.kgu.studywithme.global.dto.ResponseWrapper;
 import io.swagger.v3.oas.annotations.Operation;
