+++ src/test/java/com/kgu/studywithme/auth/domain/oauth/OAuthProviderTest.java
@@ -1,13 +1,13 @@
-package com.kgu.studywithme.auth.utils;
+package com.kgu.studywithme.auth.domain.oauth;
 
 import com.kgu.studywithme.auth.exception.AuthErrorCode;
 import com.kgu.studywithme.global.exception.StudyWithMeException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static com.kgu.studywithme.auth.utils.OAuthProvider.GOOGLE;
-import static com.kgu.studywithme.auth.utils.OAuthProvider.KAKAO;
-import static com.kgu.studywithme.auth.utils.OAuthProvider.NAVER;
+import static com.kgu.studywithme.auth.domain.oauth.OAuthProvider.GOOGLE;
+import static com.kgu.studywithme.auth.domain.oauth.OAuthProvider.KAKAO;
+import static com.kgu.studywithme.auth.domain.oauth.OAuthProvider.NAVER;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
