+++ src/test/java/com/kgu/studywithme/common/stub/StubOAuthConnector.java
@@ -1,12 +1,12 @@
 package com.kgu.studywithme.common.stub;
 
 import com.kgu.studywithme.auth.application.adapter.OAuthConnector;
+import com.kgu.studywithme.auth.domain.oauth.OAuthProvider;
 import com.kgu.studywithme.auth.domain.oauth.OAuthTokenResponse;
 import com.kgu.studywithme.auth.domain.oauth.OAuthUserResponse;
-import com.kgu.studywithme.auth.utils.OAuthProvider;
 import com.kgu.studywithme.common.fixture.OAuthFixture;
 
-import static com.kgu.studywithme.auth.utils.OAuthProvider.GOOGLE;
+import static com.kgu.studywithme.auth.domain.oauth.OAuthProvider.GOOGLE;
 
 public class StubOAuthConnector implements OAuthConnector {
     @Override
