+++ src/test/java/com/kgu/studywithme/common/stub/StubOAuthUriGenerator.java
@@ -1,9 +1,9 @@
 package com.kgu.studywithme.common.stub;
 
 import com.kgu.studywithme.auth.application.adapter.OAuthUriGenerator;
-import com.kgu.studywithme.auth.utils.OAuthProvider;
+import com.kgu.studywithme.auth.domain.oauth.OAuthProvider;
 
-import static com.kgu.studywithme.auth.utils.OAuthProvider.GOOGLE;
+import static com.kgu.studywithme.auth.domain.oauth.OAuthProvider.GOOGLE;
 
 public class StubOAuthUriGenerator implements OAuthUriGenerator {
     @Override
